requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Bilibili Processing
yt-dlp>=2023.11.0
openai-whisper>=20231117

# LLM
openai>=1.3.0
anthropic>=0.8.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utils
python-dotenv>=1.0.0
